Update patch set 15
Patch Set 15:

The settings.py code has been changed to check for the existence of a custom settings file at a predefined location. It emits a log warning if this custom settings file is not found.

However if the custom file is not found it will rely entirely on the default settings in defaults.py and these settings are insufficient to run the application.

The behaviour for the default settings needs to be defined.

Options include:
1. running with summary data.
2. running as an dummy app that just displays a web page with a message that the app is not functional.

Patch-set: 15
Label: Verified=0