Update patch set 15
Patch Set 15:

> The settings.py code has been changed to check for the existence of
 > a custom settings file at a predefined location. It emits a log
 > warning if this custom settings file is not found.
 > 
 > However if the custom file is not found it will rely entirely on
 > the default settings in defaults.py and these settings are
 > insufficient to run the application.
 > 
 > The behaviour for the default settings needs to be defined.
 > 
 > Options include:
 > 1. running with summary data.
 > 2. running as an dummy app that just displays a web page with a
 > message that the app is not functional.

1. what is summary data?

2. no

I think we don't need any data, the pages should work when there are is no data.  it would be nice if there was some fixture we could load into the database that contained say very small sample of allocations, say 5, just enough so you could actually test the application without needing the whole allocations DB.

i also noticed that the url isn't in the navigation bar? is that on purpose?

Patch-set: 15