Update patch set 16
Patch Set 16:

> I meant to say dummy data. Got caught by autocorrect.
 > 
 > At present navigating to the index page causes a failure as the
 > code attempts to load from nagios which is not accessible or the
 > dummy nagios files I use on UAT which don't make it into the git
 > repo.

I think a really small amount of dummy data is required for the allocations, i can't easily test your changes without data, i would prefer not to have to use production DB snapshots.

I'll have a look at fixing the status page so if nagios is down it still works.

 > Not sure what's meant by the missing URL.

There is no way to navigate to the allocations page, the only way to get to the page seems to be by typing in the url.

Patch-set: 16