Update patch set 16
Patch Set 16:

Re:
======
why not do
 email_domain.split('.', 1)[0] in ('my', 'ems', 'exchange', 'groupwise')
======
I have changed this, but it broke the tests.
I have tweaked the code so that it passes.

It should be understood that these kinds of changes are risky as they are
an entrance point for bugs. The unit test coverage is not good enough 
to keep such bugs at bay.

Patch-set: 16