Fixed error views
So when Django is not in debug mode and sees a 500 response from any
view.  It will catch it and replace it with the generic error page.

It also seems that in Django 1.6 the handler400 is responsible for
catching many of the errors.  Setting this to our error 500 page will
result in more errors being actually skinned and having navigation.

Change-Id: I09c765d4a5d49c621bb3fcc622bd8b5a8639d2cd