Update patch set 4
Patch Set 4:

Re: I also noticed that you can "drill down" selecting the top level category over and over again. is this the desired behaviour?

Yes. It seems some applicants don't use a fine grained categorisation of their FOR. So it was decided to use a dummy FOR4 and FOR6 based on their FOR2. This ensures some degree of consistency.

Patch-set: 4