Fixed allocations wheel display on mobile devices
Before this change the height of the wheel would match the height of the
table.  So often when the table is larger than the wheel you would need
to scroll through whitespace to find the table.  This patch also fixes
the allocation resource selector so that the buttons are displayed
bedside each other.

Change-Id: Ic9ffc9cac061cc30b7af1ab07363ddfe1f569590