Update patch set 1
Patch Set 1:

No need to kill this review.

You can download the other review as a dependency with `git-review -d <review-id>`, note the branch it creates, then do `git-review -d 1917` to download this change, then rebase this change onto the dependent branch and do a git-review. It will tell you it plans to push 2 commits, but since the dependent commit hasn't changed it will only create a new patch set for this particular review.

Patch-set: 1