Upgrade some graphs to D3v4 and add some style enhancements
D3v3 is quite old, so this commit starts to upgrade to v4 and we
may upgrade further. The exception is the growth graph which uses
the NV library that only works with D3v3, so that is left as is.

Also, we apply some style fixes and a new colour palette for
the allocations graph

Change-Id: Ib67d6109477737280f5fce7d2325dd8f60ba7b22