Add setting to enable serving /static
... in non-debug mode.   It is not the ideal to serve static content
from Django(!), but it will do until we have implemented an Apache
front end.  Baby steps ...

(Serving static content out of Swift doesn't work because of problems
induced by a limitation of django-swift-storage and its interaction
with compression of static files.  Compression is now necessary because
we have switched to using scss instead of css.)

Change-Id: I8ff17a5f1653bf7e4ac950e6e599a2bd0b9deb9b