Selenium tests for Outages
Selenium tests for the outage views and the workflows for
creating, starting, updating, completing, reopening and cancelling
scheduled and unscheduled outages.

This requires a bunch of changes to tox.ini, setup and so forth
so that we can run tests using "django-admin test"

Also, some changes to templates to add id's (for testing), correct
titles and other tweaks.

Change-Id: I35a37e6e62de0391e07b726184a6de4e4fa88cbb