Fix bug when unscheduled outage has no update
If an operator creates an unscheduled outage but doesn't
create an outage update it will lead to 500's on list
and details pages due to some missing logic in disaplaying
the status. TO fix this we'll set the status to "Investigating"
if the outage doesn't have an update which is the default
status for an unscheduled outage.

We also need a default for the severity which has be set to
SIGNIFICANT

Change-Id: I82bde501e7f0cfd2d89ad60c1e17f02862c2fc09