Add outage lists to front page
Add active, recently completed and upcoming outage lists
to bottom of Langstroth front page ... per Darcelle's design.

Also, refactor the templates so that the same html is used
for the summary lists here as in the all announcements page.

TODO:

There is a problem with the alignment of the "View announcement" links, etc

Change-Id: I46a5bdecc9a5b88b5d962c1bfb583b23f8c840e1