Fix Nagios caching bugs
The cached availability is supposed to expire after 10 minutes.
The previous logic would reset the timer if there was a cache
hit.  That meant that the availability shown by the status site
could get progressively staler ... if the site was being polled
frequently enough.

A similar problem happens with the cached status.  A second problem
is that the cache timeout for the status (10 minutes) was too long.
Changed to 1 minute.

Finally, refined the exception handling so that we log exceptions
from the Nagios interactions.

Change-Id: I8c8a6df531e3df9c023a278ae87e47e69f61013d