Fix discrepancy in site availablity
It was noticed that the site availability percentages did
not add up to 100%.  It turns out that we dropping the "unknown"
availability; i.e. cases where the site compute tempest tests
were not run; e.g. due to a problem with the tests themselves
or with Jenkins.

This change treats "unknown" as "ok"

Change-Id: I40c22a7d4e4ef073b9e23030fd95b83e599cf025